refactor: se realizaron cambios en pom, se agregró plugin y dependencias de OpenApi, se cambiaron controladores, modelos, repositorios y servicios para que se pueda generar las especificaciones de apis en base al archivo YAML, tambien se hicieron cambios leves al YAML

+++ componenteA/src/main/java/uk/techiegt/clipe/componenteA/entities/Pedido.java
@@ -6,9 +6,9 @@ import lombok.*;
 @Entity @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
 public class Pedido {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
-    private Long clienteId;
+    private Integer clienteId;
 
     /** Total calculado usando la librería del componente C */
     private double total;
