+++ componenteA/src/main/java/uk/techiegt/clipe/componenteA/repositories/ClienteRepository.java
@@ -3,4 +3,4 @@ package uk.techiegt.clipe.componenteA.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uk.techiegt.clipe.componenteA.entities.Cliente;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {}
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {}
