+++ componenteA/src/main/java/uk/techiegt/clipe/componenteA/repositories/PedidoRepository.java
@@ -3,4 +3,4 @@ package uk.techiegt.clipe.componenteA.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uk.techiegt.clipe.componenteA.entities.Pedido;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {}
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {}
